requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# NLP
transformers==4.37.0
torch==2.2.0
sentence-transformers==2.3.1
spacy==3.7.2

# Vector DB
faiss-cpu==1.8.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Guardrails
better-profanity==0.7.0

# Voice
openai-whisper
edge-tts==6.1.9
pydub==0.25.1
python-multipart==0.0.6
